requirements.txt
alembic==1.14.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
colorama==0.4.6
email-validator==2.3.0
fastapi==0.123.5
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
asyncpg==0.31.0
Telethon~=1.42.0
pydantic-settings~=2.12.0
python-jose~=3.5.0
pip~=25.1.1
rsa~=4.9.1
pyasn1~=0.6.1
typing_extensions~=4.15.0
pydantic_core~=2.41.5
pyaes~=1.6.1